+++ src/Database_Controller/City_Builder/SmartCityBuilder.java
@@ -1,4 +1,4 @@
-package Database_Controller;
+package Database_Controller.City_Builder;
 public class SmartCityBuilder implements CityBuilder {
     private City city;
