+++ src/Unit_Tests/CityBuilderTest.java
@@ -16,10 +16,10 @@ import City_Parts.Roads;
 import City_Parts.School;
 import City_Parts.Structure;
 import Database_Controller.CRUD;
-import Database_Controller.City;
-import Database_Controller.CityBuilder;
-import Database_Controller.CityDirector;
-import Database_Controller.SmartCityBuilder;
+import Database_Controller.City_Builder.City;
+import Database_Controller.City_Builder.CityBuilder;
+import Database_Controller.City_Builder.CityDirector;
+import Database_Controller.City_Builder.SmartCityBuilder;
 import FileReadWrite.CSFile;
 import FileReadWrite.IFile;
